usb: typec: ucsi: Skip setting orientation for UCSI version 2.0 and above
In case of UCSI version 2.0 and above, if the orientation is set from
glink as well, it will trigger the consumers along the graph (PHYs,
repeaters and so on) to reconfigure a second time. This might break
the consumer drivers which aren't currently implemented to drop the
second request of setting the same orientation.

So lets leave the orientation setting to the UCSI generic implementation
for all platform that implement UCSI spec 2.0 and above.

Reviewed-by: Neil Armstrong <neil.armstrong@linaro.org>
Link: https://lore.kernel.org/r/20251028-ucsi-glink-skip-orientation-for-ucsi-v2-v1-1-93e936ac59b4@linaro.org
Signed-off-by: Abel Vesa <abel.vesa@linaro.org>