dt-bindings: remoteproc: qcom,sm8550-pas: Add Kaanapali CDSP
Add remote processor PAS loader for Kaanapali CDSP processor, compatible
with earlier SM8550 with minor difference: one more sixth "shutdown-ack"
interrupt. It is not compatible with SM8650 because one memory region
"global_sync_mem" is not managed by kernel on Kaanapali so it is removed
in the remoteproc cdsp node.

Signed-off-by: Jingyi Wang <jingyi.wang@oss.qualcomm.com>
Link: https://lore.kernel.org/r/20251029-knp-remoteproc-v2-2-6c81993b52ea@oss.qualcomm.com
Signed-off-by: Abel Vesa <abel.vesa@linaro.org>